Trim btlFormation 002c4a90 raw call width

+++ src/Event/Fcl/fclCombineList.c
@@ -3240,7 +3240,7 @@ void FUN_003e0780(void)
 
   int iVar1;
 
-  u64 uVar2;
+  u32 uVar2;
